# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MCP Forge API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/specs": {
      "post": {
        "tags": [
          "Specs"
        ],
        "summary": "Create Spec Version",
        "description": "Ingest a new specification version (via upload, paste, or link), preserving matching operation selections.",
        "operationId": "create_spec_version_api_projects__project_id__specs_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Specs"
        ],
        "summary": "List Spec Versions",
        "description": "List all specification versions for a project.",
        "operationId": "list_spec_versions_api_projects__project_id__specs_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpecVersionResponse"
                  },
                  "title": "Response List Spec Versions Api Projects  Project Id  Specs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/specs/{vid}": {
      "get": {
        "tags": [
          "Specs"
        ],
        "summary": "Get Spec Version",
        "description": "Get a specific spec version including raw text.",
        "operationId": "get_spec_version_api_projects__project_id__specs__vid__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "vid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/specs/{vid}/diff": {
      "get": {
        "tags": [
          "Specs"
        ],
        "summary": "Diff Spec Version",
        "description": "Compare this spec version against another version.",
        "operationId": "diff_spec_version_api_projects__project_id__specs__vid__diff_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "vid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vid"
            }
          },
          {
            "name": "against",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target version ID or version number to diff against",
              "title": "Against"
            },
            "description": "Target version ID or version number to diff against"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecDiffReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{project_id}/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get Project Settings",
        "description": "Retrieve settings for a project.",
        "operationId": "get_project_settings_api_projects__project_id__settings_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update Project Settings",
        "description": "Update settings for a project.",
        "operationId": "update_project_settings_api_projects__project_id__settings_put",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectSettingsUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperationDiff": {
        "properties": {
          "operation_key": {
            "type": "string",
            "title": "Operation Key"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "diff_type": {
            "type": "string",
            "title": "Diff Type"
          },
          "changes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Changes"
          }
        },
        "type": "object",
        "required": [
          "operation_key",
          "method",
          "path",
          "diff_type"
        ],
        "title": "OperationDiff",
        "description": "Diff summary of an operation between two specifications."
      },
...
      "ProjectSettingsResponse": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "timeout_s": {
            "type": "integer",
            "title": "Timeout S",
            "default": 30
          },
          "max_response_chars": {
            "type": "integer",
            "title": "Max Response Chars",
            "default": 20000
          },
          "retry_safe_requests": {
            "type": "boolean",
            "title": "Retry Safe Requests",
            "default": true
          },
          "max_retries": {
            "type": "integer",
            "title": "Max Retries",
            "default": 2
          },
          "naming_style": {
            "type": "string",
            "title": "Naming Style",
            "default": "snake"
          },
          "include_writes_default": {
            "type": "boolean",
            "title": "Include Writes Default",
            "default": false
          },
          "auth_mapping": {
            "additionalProperties": true,
            "type": "object",
            "title": "Auth Mapping"
          },
          "transports": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Transports"
          },
          "tool_prefix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Prefix"
          }
        },
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "ProjectSettingsResponse"
      },
      "ProjectSettingsUpdateRequest": {
        "properties": {
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "timeout_s": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 300.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout S"
          },
          "max_response_chars": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 500000.0,
                "minimum": 100.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Response Chars"
          },
          "retry_safe_requests": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Safe Requests"
          },
          "max_retries": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Retries"
          },
          "naming_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Naming Style"
          },
          "include_writes_default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include Writes Default"
          },
          "auth_mapping": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Mapping"
          },
          "transports": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transports"
          },
          "tool_prefix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Prefix"
          }
        },
        "type": "object",
        "title": "ProjectSettingsUpdateRequest"
      },
...
      "SpecCreateRequest": {
        "properties": {
          "source_type": {
            "type": "string",
            "enum": [
              "upload",
              "paste",
              "link"
            ],
            "title": "Source Type",
            "default": "paste"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          }
        },
        "type": "object",
        "title": "SpecCreateRequest"
      },
      "SpecDiffReport": {
        "properties": {
          "base_version_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Version No"
          },
          "target_version_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Version No"
          },
          "added_count": {
            "type": "integer",
            "title": "Added Count",
            "default": 0
          },
          "removed_count": {
            "type": "integer",
            "title": "Removed Count",
            "default": 0
          },
          "modified_count": {
            "type": "integer",
            "title": "Modified Count",
            "default": 0
          },
          "unchanged_count": {
            "type": "integer",
            "title": "Unchanged Count",
            "default": 0
          },
          "operations": {
            "items": {
              "$ref": "#/components/schemas/OperationDiff"
            },
            "type": "array",
            "title": "Operations"
          },
          "title_changed": {
            "type": "boolean",
            "title": "Title Changed",
            "default": false
          },
          "version_changed": {
            "type": "boolean",
            "title": "Version Changed",
            "default": false
          },
          "servers_changed": {
            "type": "boolean",
            "title": "Servers Changed",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "base_version_no",
          "target_version_no"
        ],
        "title": "SpecDiffReport",
        "description": "Complete diff report between two spec versions."
      },
      "SpecVersionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "version_no": {
            "type": "integer",
            "title": "Version No"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "source_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "spec_kind": {
            "type": "string",
            "title": "Spec Kind"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "operation_count": {
            "type": "integer",
            "title": "Operation Count"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "version_no",
          "source_type",
          "format",
          "spec_kind",
          "sha256",
          "operation_count",
          "created_at"
        ],
        "title": "SpecVersionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}